test: move PasswordVerifyTest to isolated suite
